requirements.txt
# Legal Ops Platform Dependencies
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
python-magic==0.4.27
Pillow==10.1.0

# Email
fastapi-mail==1.4.1

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# Environment
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# JSON handling
orjson==3.9.10

# Caching
aiocache==0.12.2

# AI/ML (for future AI features)
openai==1.3.7
langchain==0.0.340

# PDF generation
reportlab==4.0.7
weasyprint==60.2

# Excel/CSV handling
openpyxl==3.1.2
pandas==2.1.4

# API documentation
mkdocs==1.5.3
mkdocs-material==9.4.8
